requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
decorator==5.2.1
defusedxml==0.7.1
exceptiongroup==1.2.2
executing==2.2.0
fastmcp==2.2.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipython==9.2.0
ipython-pygments-lexers==1.1.1
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.1.0
jedi==0.19.2
jira==3.8.0
keyring==25.6.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mcp==1.7.1
mdurl==0.1.2
more-itertools==10.7.0
oauthlib==3.2.2
openapi-pydantic==0.5.1
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
prompt-toolkit==3.0.51
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.11.4
pydantic-core==2.33.2
pydantic-settings==2.9.1
pygments==2.19.1
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==2.3.4
stack-data==0.6.3
starlette==0.46.2
traitlets==5.14.3
typer==0.15.3
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
websockets==15.0.1
